RuntimeUtility Cleanup
